+++ source/descriptor_heap.h
@@ -9,7 +9,7 @@ namespace gfx {
     struct Device;
 
     struct DescriptorHeap {
-        DescriptorHeap(const Device& device, D3D12_DESCRIPTOR_HEAP_TYPE type, D3D12_DESCRIPTOR_HEAP_FLAGS flags, size_t n_descriptors);
+        DescriptorHeap(const Device& device, D3D12_DESCRIPTOR_HEAP_TYPE type, D3D12_DESCRIPTOR_HEAP_FLAGS flags, uint32_t n_descriptors);
         ResourceHandle alloc_descriptor(ResourceType type);
         void free_descriptor(ResourceHandle id);
         D3D12_CPU_DESCRIPTOR_HANDLE fetch_cpu_handle(const ResourceHandle& id);
